fix: remove dashboard section
