+++ components/DashboardScreen[stash]/dashboard.tsx
@@ -1,0 +1,246 @@
+/**
+* This code was generated by v0 by Vercel.
+* @see https://v0.dev/t/WsDGoiFYZZw
+* Documentation: https://v0.dev/docs#integrating-generated-code-into-your-nextjs-app
+*/
+
+/** Add fonts into your Next.js project:
+
+import { Inter } from 'next/font/google'
+
+inter({
+  subsets: ['latin'],
+  display: 'swap',
+})
+
+To read more about using these font, please visit the Next.js documentation:
+- App Directory: https://nextjs.org/docs/app/building-your-application/optimizing/fonts
+- Pages Directory: https://nextjs.org/docs/pages/building-your-application/optimizing/fonts
+**/
+
+import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card"
+import { Button } from "@/components/ui/button"
+// import { CartesianGrid, XAxis, Line, LineChart, Bar, BarChart } from "recharts"
+// import { ChartTooltipContent, ChartTooltip, ChartContainer } from "@/components/ui/chart"
+
+export function Dashboard() {
+  return (
+    <main className="flex-1 p-8">
+      <div className="max-w-4xl mx-auto">
+        <h2 className="text-2xl font-bold mb-4">Dashboard</h2>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
+          {/* Weekly overview */}
+          <Card>
+            <CardHeader>
+              <CardTitle>Weekly Overview</CardTitle>
+            </CardHeader>
+            <CardContent className="flex flex-col gap-4">
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    📈
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Tasks Completed</h3>
+                    <p className="text-sm text-[#8b949e]">12 this week</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    ⏱️
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Total Time Tracked</h3>
+                    <p className="text-sm text-[#8b949e]">32 hours this week</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    🔥
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Streak</h3>
+                    <p className="text-sm text-[#8b949e]">5 days</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+            </CardContent>
+          </Card>
+
+          {/* Productivity insights */}
+          <Card>
+            <CardHeader>
+              <CardTitle>Productivity Insights</CardTitle>
+            </CardHeader>
+            <CardContent className="flex flex-col gap-4">
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    📊
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Focus Time</h3>
+                    <p className="text-sm text-[#8b949e]">75% of work time</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    🕰️
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Breaks Taken</h3>
+                    <p className="text-sm text-[#8b949e]">5 breaks this week</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+              <div className="flex items-center justify-between">
+                <div className="flex items-center gap-2">
+                  <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
+                    📆
+                  </div>
+                  <div>
+                    <h3 className="text-lg font-bold">Upcoming Tasks</h3>
+                    <p className="text-sm text-[#8b949e]">3 tasks due this week</p>
+                  </div>
+                </div>
+                <Button variant="secondary" size="sm">
+                  View
+                </Button>
+              </div>
+            </CardContent>
+          </Card>
+        </div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-8 mt-8">
+
+
+          {/* Completed tasks trends*/}
+          <Card>
+            <CardHeader>
+              <CardTitle>Task Completion Trend</CardTitle>
+            </CardHeader>
+            <CardContent>
+              {/* <LinechartChart className="aspect-[16/9]" /> */}
+            </CardContent>
+          </Card>
+
+          {/* Time tracked by project */}
+          <Card>
+            <CardHeader>
+              <CardTitle>Time Tracked by Project</CardTitle>
+            </CardHeader>
+            <CardContent>
+              {/* <BarchartChart className="aspect-[16/9]" /> */}
+            </CardContent>
+          </Card>
+
+          {/* More metrics should be added here */}
+        </div>
+      </div>
+    </main>
+  )
+}
+
+// function BarchartChart(props: React.HTMLAttributes<HTMLDivElement>) {
+//   return (
+//     <div {...props}>
+//       <ChartContainer
+//         config={{
+//           desktop: {
+//             label: "Desktop",
+//             color: "hsl(var(--chart-1))",
+//           },
+//         }}
+//         className="min-h-[300px]"
+//       >
+//         <BarChart
+//           accessibilityLayer
+//           data={[
+//             { month: "January", desktop: 186 },
+//             { month: "February", desktop: 305 },
+//             { month: "March", desktop: 237 },
+//             { month: "April", desktop: 73 },
+//             { month: "May", desktop: 209 },
+//             { month: "June", desktop: 214 },
+//           ]}
+//         >
+//           <CartesianGrid vertical={false} />
+//           <XAxis
+//             dataKey="month"
+//             tickLine={false}
+//             tickMargin={10}
+//             axisLine={false}
+//             tickFormatter={(value) => value.slice(0, 3)}
+//           />
+//           <ChartTooltip cursor={false} content={<ChartTooltipContent hideLabel />} />
+//           <Bar dataKey="desktop" fill="var(--color-desktop)" radius={8} />
+//         </BarChart>
+//       </ChartContainer>
+//     </div>
+//   )
+// }
+
+// function LinechartChart(props: React.HTMLAttributes<HTMLDivElement>) {
+//   return (
+//     <div {...props}>
+//       <ChartContainer
+//         config={{
+//           desktop: {
+//             label: "Desktop",
+//             color: "hsl(var(--chart-1))",
+//           },
+//         }}
+//       >
+//         <LineChart
+//           accessibilityLayer
+//           data={[
+//             { month: "January", desktop: 186 },
+//             { month: "February", desktop: 305 },
+//             { month: "March", desktop: 237 },
+//             { month: "April", desktop: 73 },
+//             { month: "May", desktop: 209 },
+//             { month: "June", desktop: 214 },
+//           ]}
+//           margin={{
+//             left: 12,
+//             right: 12,
+//           }}
+//         >
+//           <CartesianGrid vertical={false} />
+//           <XAxis
+//             dataKey="month"
+//             tickLine={false}
+//             axisLine={false}
+//             tickMargin={8}
+//             tickFormatter={(value) => value.slice(0, 3)}
+//           />
+//           <ChartTooltip cursor={false} content={<ChartTooltipContent hideLabel />} />
+//           <Line dataKey="desktop" type="natural" stroke="var(--color-desktop)" strokeWidth={2} dot={false} />
+//         </LineChart>
+//       </ChartContainer>
+//     </div>
+//   )
+// }
+
