--- components/dashboard.tsx
@@ -1,331 +1,0 @@
-/**
-* This code was generated by v0 by Vercel.
-* @see https://v0.dev/t/WsDGoiFYZZw
-* Documentation: https://v0.dev/docs#integrating-generated-code-into-your-nextjs-app
-*/
-
-/** Add fonts into your Next.js project:
-
-import { Inter } from 'next/font/google'
-
-inter({
-  subsets: ['latin'],
-  display: 'swap',
-})
-
-To read more about using these font, please visit the Next.js documentation:
-- App Directory: https://nextjs.org/docs/app/building-your-application/optimizing/fonts
-- Pages Directory: https://nextjs.org/docs/pages/building-your-application/optimizing/fonts
-**/
-import Link from "next/link"
-import Image from "next/image"
-import { DropdownMenu, DropdownMenuTrigger, DropdownMenuContent, DropdownMenuItem } from "@/components/ui/dropdown-menu"
-import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar"
-import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card"
-import { Button } from "@/components/ui/button"
-// import { CartesianGrid, XAxis, Line, LineChart, Bar, BarChart } from "recharts"
-import { ChartTooltipContent, ChartTooltip, ChartContainer } from "@/components/ui/chart"
-
-export function Dashboard() {
-  return (
-    <div className="flex flex-col h-screen bg-[#0d1117] text-white">
-      <header className="bg-[#161b22] py-4 px-6 flex items-center justify-between">
-        <div className="flex items-center gap-4">
-          <Link href="#" prefetch={false}>
-            <RocketIcon className="h-6 w-6 text-[#58a6ff]" />
-          </Link>
-          <h1 className="text-xl font-bold">Timekeeper</h1>
-        </div>
-        <div className="flex items-center gap-4">
-          <Link href="#" className="hover:text-[#58a6ff]" prefetch={false}>
-            Tasks
-          </Link>
-          <Link href="#" className="hover:text-[#58a6ff]" prefetch={false}>
-            Logs
-          </Link>
-          <Link href="#" className="hover:text-[#58a6ff]" prefetch={false}>
-            Dashboard
-          </Link>
-          <div className="flex items-center gap-2">
-          <Image src="/placeholder.svg" alt="User Avatar" width={32} height={32} className="rounded-full" />
-            <span>
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Avatar className="h-8 w-8">
-                    <AvatarImage src="/placeholder-user.jpg" />
-                    <AvatarFallback>JD</AvatarFallback>
-                  </Avatar>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent align="end">
-                  <DropdownMenuItem>
-                    <Link href="#" className="flex items-center gap-2" prefetch={false}>
-                      <div className="h-4 w-4" />
-                      <span>Profile</span>
-                    </Link>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem>
-                    <Link href="#" className="flex items-center gap-2" prefetch={false}>
-                      <div className="h-4 w-4" />
-                      <span>Settings</span>
-                    </Link>
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </span>
-          </div>
-        </div>
-      </header>
-      <main className="flex-1 p-8">
-        <div className="max-w-4xl mx-auto">
-          <h2 className="text-2xl font-bold mb-4">Dashboard</h2>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-8">
-            <Card>
-              <CardHeader>
-                <CardTitle>Weekly Overview</CardTitle>
-              </CardHeader>
-              <CardContent className="flex flex-col gap-4">
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      📈
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Tasks Completed</h3>
-                      <p className="text-sm text-[#8b949e]">12 this week</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      ⏱️
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Total Time Tracked</h3>
-                      <p className="text-sm text-[#8b949e]">32 hours this week</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      🔥
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Streak</h3>
-                      <p className="text-sm text-[#8b949e]">5 days</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-              </CardContent>
-            </Card>
-            <Card>
-              <CardHeader>
-                <CardTitle>Productivity Insights</CardTitle>
-              </CardHeader>
-              <CardContent className="flex flex-col gap-4">
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      📊
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Focus Time</h3>
-                      <p className="text-sm text-[#8b949e]">75% of work time</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      🕰️
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Breaks Taken</h3>
-                      <p className="text-sm text-[#8b949e]">5 breaks this week</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#0a3069] rounded-full w-10 h-10 flex items-center justify-center text-2xl">
-                      📆
-                    </div>
-                    <div>
-                      <h3 className="text-lg font-bold">Upcoming Tasks</h3>
-                      <p className="text-sm text-[#8b949e]">3 tasks due this week</p>
-                    </div>
-                  </div>
-                  <Button variant="secondary" size="sm">
-                    View
-                  </Button>
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-8 mt-8">
-            <Card>
-              <CardHeader>
-                <CardTitle>Task Completion Trend</CardTitle>
-              </CardHeader>
-              <CardContent>
-                {/* <LinechartChart className="aspect-[16/9]" /> */}
-              </CardContent>
-            </Card>
-            <Card>
-              <CardHeader>
-                <CardTitle>Time Tracked by Project</CardTitle>
-              </CardHeader>
-              <CardContent>
-                {/* <BarchartChart className="aspect-[16/9]" /> */}
-              </CardContent>
-            </Card>
-          </div>
-        </div>
-      </main>
-      <footer className="bg-[#161b22] py-4 px-6 text-center text-[#8b949e]">
-        &copy; 2023 Timekeeper. All rights reserved.
-      </footer>
-    </div>
-  )
-}
-
-function BarchartChart(props: React.HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div {...props}>
-      <ChartContainer
-        config={{
-          desktop: {
-            label: "Desktop",
-            color: "hsl(var(--chart-1))",
-          },
-        }}
-        className="min-h-[300px]"
-      >
-        <BarChart
-          accessibilityLayer
-          data={[
-            { month: "January", desktop: 186 },
-            { month: "February", desktop: 305 },
-            { month: "March", desktop: 237 },
-            { month: "April", desktop: 73 },
-            { month: "May", desktop: 209 },
-            { month: "June", desktop: 214 },
-          ]}
-        >
-          <CartesianGrid vertical={false} />
-          <XAxis
-            dataKey="month"
-            tickLine={false}
-            tickMargin={10}
-            axisLine={false}
-            tickFormatter={(value) => value.slice(0, 3)}
-          />
-          <ChartTooltip cursor={false} content={<ChartTooltipContent hideLabel />} />
-          <Bar dataKey="desktop" fill="var(--color-desktop)" radius={8} />
-        </BarChart>
-      </ChartContainer>
-    </div>
-  )
-}
-
-function LinechartChart(props: React.HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div {...props}>
-      <ChartContainer
-        config={{
-          desktop: {
-            label: "Desktop",
-            color: "hsl(var(--chart-1))",
-          },
-        }}
-      >
-        <LineChart
-          accessibilityLayer
-          data={[
-            { month: "January", desktop: 186 },
-            { month: "February", desktop: 305 },
-            { month: "March", desktop: 237 },
-            { month: "April", desktop: 73 },
-            { month: "May", desktop: 209 },
-            { month: "June", desktop: 214 },
-          ]}
-          margin={{
-            left: 12,
-            right: 12,
-          }}
-        >
-          <CartesianGrid vertical={false} />
-          <XAxis
-            dataKey="month"
-            tickLine={false}
-            axisLine={false}
-            tickMargin={8}
-            tickFormatter={(value) => value.slice(0, 3)}
-          />
-          <ChartTooltip cursor={false} content={<ChartTooltipContent hideLabel />} />
-          <Line dataKey="desktop" type="natural" stroke="var(--color-desktop)" strokeWidth={2} dot={false} />
-        </LineChart>
-      </ChartContainer>
-    </div>
-  )
-}
-
-
-function RocketIcon(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M4.5 16.5c-1.5 1.26-2 5-2 5s3.74-.5 5-2c.71-.84.7-2.13-.09-2.91a2.18 2.18 0 0 0-2.91-.09z" />
-      <path d="m12 15-3-3a22 22 0 0 1 2-3.95A12.88 12.88 0 0 1 22 2c0 2.72-.78 7.5-6 11a22.35 22.35 0 0 1-4 2z" />
-      <path d="M9 12H4s.55-3.03 2-4c1.62-1.08 5 0 5 0" />
-      <path d="M12 15v5s3.03-.55 4-2c1.08-1.62 0-5 0-5" />
-    </svg>
-  )
-}
-
-
-function XIcon(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M18 6 6 18" />
-      <path d="m6 6 12 12" />
-    </svg>
-  )
-}
